# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Mi API nueva",
        "description": "DescripciÃ³n de mi API.",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/pagos/user/{user_id}": {
            "get": {
                "tags": [
                    "Pagos"
                ],
                "summary": "Listar pagos de un usuario especÃ­fico con paginaciÃ³n",
                "operationId": "bfd6a6d58c4a77c909578f7f64df5a83",
                "parameters": [
                    {
                        "name": "user_id",
                        "in": "path",
                        "description": "ID del usuario",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "NÃºmero de pÃ¡gina de la paginaciÃ³n (por defecto 1)",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "Cantidad de registros por pÃ¡gina (por defecto 10)",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 10
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista paginada de pagos del usuario",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "type": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Pagos del usuario obtenidos correctamente"
                                        },
                                        "data": {
                                            "properties": {
                                                "current_page": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "data": {
                                                    "type": "array",
                                                    "items": {
                                                        "properties": {
                                                            "id": {
                                                                "type": "integer",
                                                                "example": 1
                                                            },
                                                            "user_id": {
                                                                "type": "integer",
                                                                "example": 2
                                                            },
                                                            "concepto": {
                                                                "type": "string",
                                                                "example": "Cuota mantenimiento"
                                                            },
                                                            "fecha": {
                                                                "type": "string",
                                                                "example": "2024-02-10"
                                                            },
                                                            "vence": {
                                                                "type": "string",
                                                                "example": "2024-02-20"
                                                            },
                                                            "documento": {
                                                                "type": "string",
                                                                "example": "03/B002-0058037"
                                                            },
                                                            "currency_id": {
                                                                "type": "integer",
                                                                "example": 1
                                                            },
                                                            "importe": {
                                                                "type": "number",
                                                                "format": "float",
                                                                "example": 290
                                                            },
                                                            "estado": {
                                                                "type": "string",
                                                                "example": "Pagado"
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                },
                                                "total": {
                                                    "type": "integer",
                                                    "example": 30
                                                },
                                                "per_page": {
                                                    "type": "integer",
                                                    "example": 10
                                                },
                                                "last_page": {
                                                    "type": "integer",
                                                    "example": 3
                                                },
                                                "from": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "to": {
                                                    "type": "integer",
                                                    "example": 10
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "Bearer",
                "scheme": "bearer"
            }
        }
    },
...
}